fix(FOROME-1274): saving preset on ws-page after reloading
